backend/requirements.txt
django==3.1.4

django-cors-headers==3.5.0
djangorestframework==3.12.2

redis==3.5.3
django-redis==4.12.1
channels==3.0.2
channels-redis==3.2.0
twisted[tls,http2]==20.3.0

pillow==8.0.1
django-cleanup==5.1.0

whitenoise==5.2.0
python-decouple==3.3
gunicorn==20.0.4

psycopg2==2.8.6
psycopg2-binary==2.8.6

mixer==6.1.3
pytest==6.1.2
pytest-cov==2.10.1
pytest-django==4.1.0
dj-inmemorystorage==2.1.0

pydantic==1.7.3

sentry-sdk==0.19.5
